admin: Remove partial initialisation of InviteCode
Just a refactoring so that the properties of the new InviteCode created
by an admin are set once (via the create constructor) rather than one at
a time.

This avoids the situation where we 'have' an InviteCode but the model
object isn't valid in its own terms since it's not fully initialised.